refactor(Discussion): Reorganize Discussion and Comments components
