+++ components/Discussion/Discussion.js
@@ -1,40 +1,8 @@
 import React, { PureComponent } from 'react'
-import { compose } from 'react-apollo'
-import { css } from 'glamor'
-import withT from '../../lib/withT'
-import { A, colors, fontStyles, mediaQueries } from '@project-r/styleguide'
-
-import { withDiscussionCommentsCount } from './graphql/enhancers/withDiscussionCommentsCount'
 
 import DiscussionCommentComposer from './DiscussionCommentComposer'
 import NotificationOptions from './NotificationOptions'
 import Comments from './Comments'
-
-const styles = {
-  orderByContainer: css({
-    margin: '20px 0'
-  }),
-  orderBy: css({
-    ...fontStyles.sansSerifRegular16,
-    outline: 'none',
-    color: colors.text,
-    WebkitAppearance: 'none',
-    background: 'transparent',
-    border: 'none',
-    padding: '0',
-    cursor: 'pointer',
-    marginRight: '20px',
-    [mediaQueries.mUp]: {
-      marginRight: '40px'
-    }
-  }),
-  selectedOrderBy: css({
-    textDecoration: 'underline'
-  }),
-  emptyDiscussion: css({
-    margin: '20px 0'
-  })
-}
 
 class Discussion extends PureComponent {
   state = {
@@ -42,13 +10,6 @@ class Discussion extends PureComponent {
      * DiscussionOrder ('DATE' | 'VOTES' | 'REPLIES')
      */
     orderBy: 'DATE',
-
-    /**
-     * This is a counter which we increment whenever we want to reset the Apollo
-     * cache and reload the comments from the server. The <Comments> component
-     * detects if this counter increments and invokes the Apollo refetch() function.
-     */
-    reload: 0,
 
     now: Date.now()
   }
@@ -63,22 +24,13 @@ class Discussion extends PureComponent {
     clearInterval(this.intervalId)
   }
 
-  onReload = e => {
-    e.preventDefault()
-    this.setState(({ reload }) => ({ reload: reload + 1 }))
+  onOrderBy = orderBy => {
+    this.setState({ orderBy })
   }
 
   render () {
-    const { t, discussionCommentsCount, discussionId, focusId = null, mute, meta, sharePath } = this.props
+    const { discussionId, focusId = null, mute, meta, sharePath } = this.props
     const { orderBy, reload, now } = this.state
-
-    const OrderBy = ({ children, value }) => (
-      <button {...styles.orderBy} {...(orderBy === value ? styles.selectedOrderBy : {})} onClick={() => {
-        this.setState({ orderBy: value })
-      }}>
-        {t(`components/Discussion/OrderBy/${value}`)}
-      </button>
-    )
 
     return (
       <div data-discussion-id={discussionId}>
@@ -93,48 +45,22 @@ class Discussion extends PureComponent {
 
         <NotificationOptions discussionId={discussionId} mute={mute} />
 
-        {(() => {
-          if (discussionCommentsCount > 0) {
-            return (
-              <>
-                <div {...styles.orderByContainer}>
-                  <OrderBy value='DATE' />
-                  <OrderBy value='VOTES' />
-                  <OrderBy value='REPLIES' />
-                  <A style={{ float: 'right', lineHeight: '25px', cursor: 'pointer' }} href='' onClick={this.onReload}>
-                    {t('components/Discussion/reload')}
-                  </A>
-                  <br style={{ clear: 'both' }} />
-                </div>
-
-                <Comments
-                  depth={1}
-                  key={orderBy}
-                  discussionId={discussionId}
-                  focusId={focusId}
-                  parentId={null}
-                  reload={reload}
-                  orderBy={orderBy}
-                  now={now}
-                  meta={meta}
-                  sharePath={sharePath}
-                />
-              </>
-            )
-          } else {
-            return <EmptyDiscussion t={t} />
-          }
-        })()}
+        <Comments
+          depth={1}
+          key={orderBy}
+          discussionId={discussionId}
+          focusId={focusId}
+          parentId={null}
+          reload={reload}
+          orderBy={orderBy}
+          now={now}
+          meta={meta}
+          sharePath={sharePath}
+          onOrderBy={this.onOrderBy}
+        />
       </div>
     )
   }
 }
 
-export default compose(
-  withT,
-  withDiscussionCommentsCount
-)(Discussion)
-
-const EmptyDiscussion = ({ t }) => (
-  <div {...styles.emptyDiscussion}>{t('components/Discussion/empty')}</div>
-)
+export default Discussion
