+++ components/Discussion/DiscussionCommentComposer.js
@@ -8,6 +8,7 @@ import produce from 'immer'
 
 import { withDiscussionDisplayAuthor } from './graphql/enhancers/withDiscussionDisplayAuthor'
 import { withDiscussionPreferences } from './graphql/enhancers/withDiscussionPreferences'
+import { withDiscussionComments } from './graphql/enhancers/withDiscussionComments'
 import { withSubmitComment } from './graphql/enhancers/withSubmitComment'
 
 import DiscussionPreferences from './DiscussionPreferences'
@@ -178,5 +179,6 @@ export default compose(
   withMe,
   withDiscussionDisplayAuthor,
   withDiscussionPreferences,
+  withDiscussionComments,
   withSubmitComment
 )(DiscussionCommentComposer)
